+++ lib/i18n/strings.g.dart
@@ -4,9 +4,9 @@
 /// To regenerate, run: `dart run slang`
 ///
 /// Locales: 2
-/// Strings: 2102 (1051 per locale)
+/// Strings: 2116 (1058 per locale)
 ///
-/// Built on 2026-08-08 at 05:15 UTC
+/// Built on 2026-08-08 at 05:27 UTC
 
 // coverage:ignore-file
 // ignore_for_file: type=lint, unused_import